core: pager: hide tee_pager_tbl_info
Hides tee_pager_tbl_info and provides new needed functions.
