+++ core/arch/arm/include/mm/tee_pager.h
@@ -33,16 +33,37 @@
 #include <kernel/panic.h>
 #include <kernel/user_ta.h>
 #include <mm/tee_mm.h>
+#include <mm/core_mmu.h>
 #include <string.h>
 #include <trace.h>
 
-/*
- * Reference to translation table used to map the virtual memory range
- * covered by the pager.
- */
-extern struct core_mmu_table_info tee_pager_tbl_info;
-
 struct tee_pager_area_head;
+
+/*
+ * tee_pager_early_init() - Perform early initialization of pager
+ *
+ * Panics if some error occurs
+ */
+void tee_pager_early_init(void);
+
+/*
+ * tee_pager_get_table_info() - Fills in table info for address mapped in
+ * translation table managed by the pager.
+ * @va:		address to look up
+ * @ti:		filled in table info
+ *
+ * Returns true if address is in the pager translation tables else false
+ */
+bool tee_pager_get_table_info(vaddr_t va, struct core_mmu_table_info *ti);
+
+/*
+ * tee_pager_phys_to_virt() - Translate physical address to virtual address
+ * looking in the pager page tables
+ * @pa:	address to translate
+ *
+ * Returns found virtual address or NULL on error
+ */
+void *tee_pager_phys_to_virt(paddr_t pa);
 
 /*
  * tee_pager_init() - Initialized the pager
